pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "google-genai>=1.59.0",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
